clang-tidy: modernize-use-using respect externC
* clang-tidy: modernize-use-using: add extern C case

Add an edge case for testing against extern C linkage.

Signed-off-by: Dan Kalowsky <dkalowsky@esri.com>

* clang-tidy: modernize-use-using respect extern C

Previous versions of the check would ignore if a typedef was set to
extern C.  This causes problems for files that have portions of code
that cross between the Cxx and C linkage spec.

When the checker encounters an entry that has C linkage, return and
ignore it as we cannot properly fix or check it.

Signed-off-by: Dan Kalowsky <dkalowsky@esri.com>

* Use Duncan's simplified method

Signed-off-by: Dan Kalowsky <dkalowsky@esri.com>