clang-tidy: Fix performance-unnecessary-value-param check for move calls
The check has cases of false positives when an expensive to copy type is being passed by value and moved. The check changes
the argument to a const reference and still keeps the move(), which is a no-op now. This commit fixes this by adding a matcher
that checks if the expensive to copy function parameter is being moved anywhere in the function body, and does nothing if
the matcher finds a match