clang-tidy: Change sequence of call to check for std::move()
This commit swaps the order in which the mutation analyzer is run to check for any mutations to
the function argument, and the function call to check if there argument is being moved. This is
done so that in case, something like this happens -

b = std::move(a);
a.set(...); //set new value for a i.e mutation

Then the checker should exit because it detected a mutation. The changed order refelcts that.

clang-tidy: Fix and format regression, unit tests for performance checker

This commit fixes the issues in the regression and unit tests added for the performance-unnecessary-value-param checker
The commit replaces unit tests to not use standard library headers like <string>, and use examples that can be compiled
and run without using the standard library.

The commit also formats the checker file and the regeression tests file with clang-format.

clang-tidy: remove newly added tests that are not needed

This commit removes two tests that were added in an earlier commit. The tests do not serve
any useful purpose to test the checker, therefore they have been removed in this commit.

clang-tidy: Remove tests and fix formatting

This commit removes some newly added tests that didn't serve any purpose in verifying
the checker. The commit also fixes the format of the file